add check if file or directory exists

+++ install.sh
@@ -5,13 +5,18 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cp $DIR/vimrc ~/.vimrc
 
 # Create directory for vim bundle
-mkdir -p ~/.vim/bundle
+if [ ! -d ~/.vim/bundle]; then
+    mkdir -p ~/.vim/bundle
+fi
 
-# Clone neobundle
-git clone git://github.com/Shougo/neobundle.vim ~/.vim/bundle/neobundle.vim
+if [ ! -d ~/.vim/bundle/neobundle.vim ]; then
+    # Clone neobundle
+    git clone git://github.com/Shougo/neobundle.vim ~/.vim/bundle/neobundle.vim
+fi
 
-# Clone solarized
-git clone https://github.com/altercation/vim-colors-solarized.git /tmp/solarized
-cp -rf /tmp/solarized/colors ~/.vim/
-rm -rf /tmp/solarized
-
+if [ ! -f ~/.vim/colors/solarized.vim ]; then
+    # Clone solarized
+    git clone https://github.com/altercation/vim-colors-solarized.git /tmp/solarized
+    cp -rf /tmp/solarized/colors/solarized.vim ~/.vim/colors
+    rm -rf /tmp/solarized
+fi
